python-products-service/requirements.txt
autopep8==1.5.6
click==7.1.2
Flask==1.1.2
Flask-SQLAlchemy==2.5.1
greenlet==1.0.0
importlib-metadata==3.10.0
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
pycodestyle==2.7.0
pymongo==3.11.3
python-dotenv==0.16.0
SQLAlchemy==1.4.4
toml==0.10.2
typing-extensions==3.7.4.3
Werkzeug==1.0.1
zipp==3.4.1
